Fix chat's show
