+++ arcade/arcade/cli/main.py
@@ -18,19 +18,16 @@ from arcade.cli.constants import DEFAULT_CLOUD_HOST, DEFAULT_ENGINE_HOST, LOCALH
 from arcade.cli.display import (
     display_arcade_chat_header,
     display_eval_results,
-    display_tool_details,
     display_tool_messages,
-    display_tools_table,
 )
 from arcade.cli.launcher import start_servers
+from arcade.cli.show import show_logic
 from arcade.cli.utils import (
     OrderCommands,
     compute_engine_base_url,
     compute_login_url,
-    create_cli_catalog,
     delete_deprecated_config_file,
     get_eval_files,
-    get_tools_from_engine,
     get_user_input,
     handle_chat_interaction,
     handle_tool_authorization,
@@ -177,38 +174,7 @@ def show(
     """
     Show the available toolkits or detailed information about a specific tool.
     """
-    try:
-        if local:
-            catalog = create_cli_catalog(toolkit=toolkit)
-            tools = [t.definition for t in list(catalog)]
-        else:
-            tools = get_tools_from_engine(host, port, force_tls, force_no_tls, toolkit)
-
-        if tool:
-            # Display detailed information for the specified tool
-            tool_def = next(
-                (
-                    t
-                    for t in tools
-                    if t.get_fully_qualified_name().name.lower() == tool.lower()
-                    or str(t.get_fully_qualified_name()).lower() == tool.lower()
-                ),
-                None,
-            )
-            if not tool_def:
-                console.print(f"❌ Tool '{tool}' not found.", style="bold red")
-                typer.Exit(code=1)
-            else:
-                display_tool_details(tool_def)
-        else:
-            # Display the list of tools as a table
-            display_tools_table(tools)
-
-    except Exception as e:
-        if debug:
-            raise
-        error_message = f"❌ Failed to list tools: {escape(str(e))}"
-        console.print(error_message, style="bold red")
+    show_logic(toolkit, tool, host, local, port, force_tls, force_no_tls, debug)
 
 
 @cli.command(help="Start Arcade Chat in the terminal", rich_help_panel="Launch")
@@ -282,7 +248,9 @@ def chat(
             # Add the input to history
             readline.add_history(user_input)
 
-            if handle_user_command(user_input, history, host, port, force_tls, force_no_tls, show):
+            if handle_user_command(
+                user_input, history, host, port, force_tls, force_no_tls, show_logic
+            ):
                 continue
 
             history.append({"role": "user", "content": user_input})
